trading_platform/requirements.txt
pydantic>=2.0
pydantic-settings>=2.0
fastapi>=0.100.0
uvicorn>=0.23.0
pytz>=2023.3
PyYAML>=6.0
httpx>=0.24.0
pytest>=7.4.0
python-dateutil>=2.8.0
jinja2>=3.1.0
